refactor: organize landing pg comps
