+++ client/src/pages/Landing/TextHoverEffectDemo.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TextHoverEffect } from "@/components/ui/text-hover-effect";
+import { TextHoverEffect } from "./text-hover-effect";
 
 export default function TextHoverEffectDemo() {
     return (
